# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title; it controls generated filenames.
  title: Api
  version: 0.1.0
  description: Public shipment tracking and logistics operations API.
servers:
  - url: /api
...
paths:
...
  /tracking/{trackingNumber}:
    get:
      operationId: getPublicTracking
      tags: [public]
      summary: Get public shipment tracking details
      parameters:
        - name: trackingNumber
          in: path
          required: true
          schema: { type: string, minLength: 4, maxLength: 64 }
      responses:
        "200":
          description: Shipment and tracking history without customer-identifying data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TrackingResult"
        "404":
          description: Tracking number not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /quotes:
    post:
      operationId: createQuote
      tags: [public]
      summary: Submit a quote request
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QuoteInput"
      responses:
        "201":
          description: Quote received
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Quote"
        "400":
          description: Invalid quote request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /contacts:
    post:
      operationId: createContact
      tags: [public]
      summary: Submit a contact inquiry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContactInput"
      responses:
        "201":
          description: Inquiry received
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contact"
  /me:
    get:
      operationId: getCurrentUser
      tags: [customer]
      summary: Get the signed-in user's profile
      security: [{ clerkSession: [] }]
      responses:
        "200":
          description: Current profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    patch:
      operationId: updateCurrentUser
      tags: [customer]
      summary: Update the signed-in user's profile
      security: [{ clerkSession: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProfileUpdate"
      responses:
        "200":
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /customer/shipments:
    get:
      operationId: getCustomerShipments
      tags: [customer]
      summary: List only the signed-in customer's shipments
      security: [{ clerkSession: [] }]
      responses:
        "200":
          description: Shipments
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Shipment" }
  /customer/quotes:
    get:
      operationId: getCustomerQuotes
      tags: [customer]
      summary: List only the signed-in customer's quote requests
      security: [{ clerkSession: [] }]
      responses:
        "200":
          description: Quote requests
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Quote" }
  /customer/notifications:
    get:
      operationId: getCustomerNotifications
      tags: [customer]
      summary: List only the signed-in user's notifications
      security: [{ clerkSession: [] }]
      responses:
        "200":
          description: Notifications
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Notification" }
  /customer/notifications/{id}/read:
    patch:
      operationId: markNotificationRead
      tags: [customer]
      summary: Mark an owned notification as read
      security: [{ clerkSession: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Updated notification
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Notification"
...
components:
  securitySchemes:
    clerkSession:
      type: http
      scheme: bearer
      description: Authenticated Clerk browser session (same-origin cookie).
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error: { type: string }
...
    TrackingEvent:
      type: object
      required: [id, source, status, description, eventTimestamp]
      properties:
        id: { type: integer }
        source: { type: string, enum: [internal, carrier] }
        sourceEventId: { type: ["string", "null"] }
        status: { type: string }
        location: { type: ["string", "null"] }
        description: { type: string }
        eventTimestamp: { type: string, format: date-time }
...
    Shipment:
      type: object
      required: [id, trackingNumber, serviceType, origin, destination, status, isDemo, createdAt, updatedAt]
      properties:
        id: { type: integer }
        trackingNumber: { type: string }
        customerId: { type: ["integer", "null"] }
        customerName: { type: ["string", "null"] }
        carrier: { type: ["string", "null"] }
        carrierTrackingNumber: { type: ["string", "null"] }
        serviceType: { type: string }
        origin: { type: string }
        destination: { type: string }
        currentLocation: { type: ["string", "null"] }
        status: { type: string }
        weight: { type: ["number", "null"] }
        dimensions: { type: ["string", "null"] }
        packageDescription: { type: ["string", "null"] }
        estimatedDelivery: { type: ["string", "null"], format: date }
        actualDelivery: { type: ["string", "null"], format: date }
        isDemo: { type: boolean }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    ShipmentDetail:
      type: object
      required: [shipment, events]
      properties:
        shipment: { $ref: "#/components/schemas/Shipment" }
        events:
          type: array
          items: { $ref: "#/components/schemas/TrackingEvent" }
    TrackingResult:
      allOf:
        - $ref: "#/components/schemas/ShipmentDetail"
...
    Quote:
      type: object
      required: [id, customerName, email, origin, destination, shipmentType, status, createdAt, updatedAt]
      properties:
        id: { type: integer }
        customerId: { type: ["integer", "null"] }
        customerName: { type: string }
        email: { type: string, format: email }
        phone: { type: ["string", "null"] }
        origin: { type: string }
        destination: { type: string }
        shipmentType: { type: string }
        weight: { type: ["number", "null"] }
        dimensions: { type: ["string", "null"] }
        description: { type: ["string", "null"] }
        preferredShippingDate: { type: ["string", "null"], format: date }
        additionalInformation: { type: ["string", "null"] }
        status: { type: string, enum: [pending, reviewing, quoted, accepted, rejected, completed] }
        adminNotes: { type: ["string", "null"] }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    QuoteInput:
      type: object
      required: [customerName, email, origin, destination, shipmentType]
      properties:
        customerName: { type: string, minLength: 2, maxLength: 120 }
        email: { type: string, format: email, maxLength: 320 }
        phone: { type: ["string", "null"], maxLength: 40 }
        origin: { type: string, minLength: 2, maxLength: 180 }
        destination: { type: string, minLength: 2, maxLength: 180 }
        shipmentType: { type: string, minLength: 2, maxLength: 80 }
        weight: { type: ["number", "null"], minimum: 0 }
        dimensions: { type: ["string", "null"], maxLength: 100 }
        description: { type: ["string", "null"], maxLength: 500 }
        preferredShippingDate: { type: ["string", "null"], format: date }
        additionalInformation: { type: ["string", "null"], maxLength: 2000 }
...
    Contact:
      type: object
      required: [id, name, email, subject, message, status, createdAt]
      properties:
        id: { type: integer }
        name: { type: string }
        email: { type: string, format: email }
        phone: { type: ["string", "null"] }
        subject: { type: string }
        message: { type: string }
        status: { type: string }
        createdAt: { type: string, format: date-time }
    ContactInput:
      type: object
      required: [name, email, subject, message]
      properties:
        name: { type: string, minLength: 2, maxLength: 120 }
        email: { type: string, format: email, maxLength: 320 }
        phone: { type: ["string", "null"], maxLength: 40 }
        subject: { type: string, minLength: 2, maxLength: 160 }
        message: { type: string, minLength: 5, maxLength: 4000 }
...
    User:
      type: object
      required: [id, clerkUserId, name, email, role, status, createdAt]
      properties:
        id: { type: integer }
        clerkUserId: { type: string }
        name: { type: string }
        email: { type: string, format: email }
        role: { type: string, enum: [super_admin, admin, staff, customer] }
        status: { type: string, enum: [active, disabled] }
        phone: { type: ["string", "null"] }
        address: { type: ["string", "null"] }
        country: { type: ["string", "null"] }
        createdAt: { type: string, format: date-time }
    ProfileUpdate:
      type: object
      properties:
        name: { type: string, minLength: 2, maxLength: 120 }
        phone: { type: ["string", "null"], maxLength: 40 }
        address: { type: ["string", "null"], maxLength: 250 }
        country: { type: ["string", "null"], maxLength: 100 }
...
    Notification:
      type: object
      required: [id, type, title, message, createdAt]
      properties:
        id: { type: integer }
        type: { type: string }
        title: { type: string }
        message: { type: string }
        readAt: { type: ["string", "null"], format: date-time }
        createdAt: { type: string, format: date-time }
...