# lib/api-zod/src/generated/types/shipmentDetail.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Public shipment tracking and logistics operations API.
 * OpenAPI spec version: 0.1.0
 */
import type { Shipment } from './shipment';
import type { TrackingEvent } from './trackingEvent';

export interface ShipmentDetail {
  shipment: Shipment;
  events: TrackingEvent[];
}