# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Public shipment tracking and logistics operations API.
 * OpenAPI spec version: 0.1.0
 */
import type { UserRole } from './userRole';
import type { UserStatus } from './userStatus';

export interface User {
  id: number;
  clerkUserId: string;
  name: string;
  email: string;
  role: UserRole;
  status: UserStatus;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  address?: string | null;
  /** @nullable */
  country?: string | null;
  createdAt: Date;
}